refactor: count 쿼리 결과가 int를 반환하도록 수정
* null인 경우 0을 반환하고 아닌 경우 int 결과를 반환
* null인 경우를 처리하기 위해
* 불필요한 메모리 사용 방

+++ src/main/java/kcs/funding/fundingboost/domain/service/FundingService.java
@@ -347,7 +347,7 @@ public class FundingService {
 
         List<MyPageFundingDetailHistoryDto> myPageFundingDetailHistoryDtos = fundings.stream()
                 .map(funding -> {
-                    Long contributors = contributorRepository.countContributorsForFunding(funding.getFundingId());
+                    int contributors = contributorRepository.countContributorsForFunding(funding.getFundingId());
                     return MyPageFundingDetailHistoryDto.fromEntity(funding, contributors);
                 })
                 .toList();
